+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- NrqlEditor - change account picker to use select element
+
 ## [1.23.6] - 2025-02-10
 
 ### Fixed